Move OID4VP request options into protocol
